Deep Neural Network/requirements.txt
astroid==2.4.2
colorama==0.4.4
cycler==0.10.0
h5py==3.1.0
isort==5.7.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.3
mccabe==0.6.1
numpy==1.19.3
Pillow==8.0.1
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
toml==0.10.2
wrapt==1.12.1
scipy==1.6.0